+++ src/AsyncSeq/AssemblyInfo.fs
@@ -1,8 +1,8 @@
-﻿namespace System
+namespace System
 open System.Reflection
 
-[<assembly: AssemblyTitleAttribute("FSharpx.Async")>]
-[<assembly: AssemblyProductAttribute("FSharpx.Async")>]
+[<assembly: AssemblyTitleAttribute("AsyncSeq")>]
+[<assembly: AssemblyProductAttribute("AsyncSeq")>]
 [<assembly: AssemblyDescriptionAttribute("Async extensions for F#")>]
 [<assembly: AssemblyVersionAttribute("1.11.1")>]
 [<assembly: AssemblyFileVersionAttribute("1.11.1")>]