+++ src/AsyncSeq/Observable.fs
@@ -21,35 +21,6 @@ type ObservableUpdate<'T> =
 
 module Observable =
 
-    /// Returns an observable that yields sliding windows of 
-    /// containing elements drawn from the input observable. 
-    /// Each window is returned as a fresh array.
-    let windowed size (input:IObservable<'T>) =
-      { new IObservable<'T[]> with
-          member x.Subscribe(observer) =
-            // Create sliding window agent for every call
-            // and redirect batches to the observer
-            let cts = new CancellationTokenSource()
-            let agent = new SlidingWindowAgent<_>(size, cts.Token)
-            agent.WindowProduced.Add(observer.OnNext)
-
-            // Subscribe to the input and send values to the agent
-            let subscription = 
-              input.Subscribe
-                ({ new IObserver<'T> with
-                    member x.OnNext(v) = agent.Enqueue(v)
-                    member x.OnCompleted() = 
-                      cts.Cancel()
-                      observer.OnCompleted()
-                    member x.OnError(e) = 
-                      cts.Cancel()
-                      observer.OnError(e) })
-
-            // Cancel subscription & cancel the agent
-            { new IDisposable with 
-                member x.Dispose() =
-                  subscription.Dispose()
-                  cts.Cancel() } }
 
     /// Creates an observable that calls the specified function (each time)
     /// after an observer is attached to the observable. This is useful to 
@@ -181,21 +152,6 @@ module Observable =
             | Choice2Of2 r when lefts.Count = 0 -> rights.pushBack(r); None
             | Choice2Of2 r -> Some(lefts.popFront(), r)
         )
-
-    let bufferWithTimeOrCount<'T> (timeSpan:TimeSpan) (count:int) (source:IObservable<'T>)=
-        let timeSpan = int timeSpan.TotalMilliseconds
-        let batch = new BatchProcessingAgent<'T>(count, timeSpan) 
-        { new IObservable<'T seq> with
-            member this.Subscribe(observer:IObserver<'T seq>) =
-                let sd = source.Subscribe(fun v -> batch.Enqueue v)
-                let dd  = batch.BatchProduced.Subscribe(fun v -> observer.OnNext v)
-                { new IDisposable with 
-                    member this.Dispose() = 
-                        sd.Dispose()
-                        dd.Dispose()
-                        (batch :> IDisposable).Dispose() 
-                }
-        }
 
     [<AbstractClass>]  
     type internal BasicObserver<'T>() =
@@ -451,10 +407,6 @@ type ObservableExtensions private () =
     [<Extension>]
     static member Delay<'TSource>(source:IObservable<'TSource>,milliseconds:int) =
         source |> Observable.delay milliseconds
-
-    [<Extension>]
-    static member BufferWithTimeOrCount<'TSource>(source:IObservable<'TSource>, timeSpan:TimeSpan, count:int) =
-        source |> Observable.bufferWithTimeOrCount timeSpan count
 
     [<Extension>]
     static member Throttle<'TSource>(source:IObservable<'TSource>, dueTime:TimeSpan) =