+++ RELEASE_NOTES.md
@@ -1,5 +1,5 @@
 ### 1.11.1 - 26.03.2015
-* BUGFIX: AsyncSeq.toBlockingSeq does not hung forever if an exception is thrown and reraise it outside - https://github.com/fsprojects/FSharpx.Async/pull/21
+* BUGFIX: AsyncSeq.toBlockingSeq does not hung forever if an exception is thrown and reraise it outside - https://github.com/fsprojects/AsyncSeq/pull/21
 * Added Async.bindChoice, Async.ParallelIgnore, AsyncSeq.zipWithAsync, AsyncSeq.zappAsync, AsyncSeq.threadStateAsync, AsyncSeq.merge, AsyncSeq.traverseOptionAsync, AsyncSeq.traverseChoiceAsync
 
 ### 1.11.0 - 27.02.2015
@@ -10,8 +10,8 @@
 * Use Paket instead of NuGet
 
 ### 1.9.9 - 23.02.2015
-* BUGFIX: AsyncSeq.skipWhile skips an extra item - https://github.com/fsprojects/FSharpx.Async/pull/2
+* BUGFIX: AsyncSeq.skipWhile skips an extra item - https://github.com/fsprojects/AsyncSeq/pull/2
  
 ### 1.9.9 - 23.02.2015
 * Copied the async helpers from FSharpx
-* BUGFIX: AsyncSeq.skipWhile skips an extra item - https://github.com/fsprojects/FSharpx.Async/pull/2
+* BUGFIX: AsyncSeq.skipWhile skips an extra item - https://github.com/fsprojects/AsyncSeq/pull/2