+++ RELEASE_NOTES.md
@@ -1,4 +1,4 @@
-### 1.12.0 - 27.03.2015
+### 1.12.1 - 27.03.2015
 * Added Async.bindChoice, Async.ParallelIgnore, AsyncSeq.zipWithAsync, AsyncSeq.zappAsync, AsyncSeq.threadStateAsync, AsyncSeq.merge, AsyncSeq.traverseOptionAsync, AsyncSeq.traverseChoiceAsync
 * Added AsyncSeq.toList, AsyncSeq.toArray, AsyncSeq.bufferByCount, AsyncSeq.unfoldAsync, AsyncSeq.concatSeq, AsyncSeq.interleave
 * Copied the AsyncSeq from FSharpx.Async